📝 Initialize Documentation on ReadTheDocs (#74)
This PR adds the initial configuration and infrastructure for building
proper project documentation on ReadTheDocs.

Building the documentation locally is as easy as
```console
nox -rs docs -- serve
```

> `nox` is best installed via `pipx install nox` (or `brew install nox`
if you are on macOS). If you don't have `pipx` (pip for applications)
you can install it with `pip install pipx` (or `brew install pipx` if
you are on macOS).

At the moment the documentation is pretty much empty. Filling it with
content will be done in a separate PR.

Signed-off-by: burgholzer <burgholzer@me.com>