DD synthesis of irreversible functions (Encoded to reversible without additional line) (#115)
## Description

This PR addresses the dd synthesis of irreversible functions which are
encoded to reversible without any need of an additional line. For the
synthesis, don't care conditions must be taken into consideration. This
PR introduces the DC node condition based on which the synthesis can be
carried forward. Appropriate tests and benchmarks are included as well.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.