Encoding irreversible function to reversible function using an additional line. (#119)
## Description

In this PR, additional functionality has been included to synthesize
encoded (using an additional line) irreversible functions. While
synthesizing, instead of p1 and p2 paths of a node, p3 and p4 paths are
to be considered as well. More details are provided in the code
documentation. Appropriate benchmarks have been included and tested.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.